netfilter: nf_conncount: optimize nf_conncount performance
From: Florian Westphal <fw@strlen.de>

This patch applies the following techniques to optimize nf_conncount
performance.

* Split tree insertion and traversal
When we have a very coarse grouping, e.g. by large subnets, zone id,
etc, it's likely that we do not need to do tree rotation because
we'll find a node where we can attach new entry.  Based on this
observation, we then make traversal lockless (tree protected
by RCU), and add extra lock in the individual nodes to protect list
insertion/deletion, thereby allowing parallel insert/delete in different
tree nodes.

* Early exit for garbage
In order to reduce gc time, we skip traversing the full list for
every node when doing garbage collection, since it is enough to zap
a couple of expired entries.

* Add garbage collection worker
Instead of doing all of garbage collection task in the packet forwarding
path, we will schedule a garbage collection worker when the number of
nodes that can be freed exceeds a threshold.

Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Yi-Hung Wei <yihung.wei@gmail.com>