netfilter: nf_conncount: Switch to plain list
From: Florian Westphal <fw@strlen.de>

This patch switches from hlist to plain list to store the list of
connections with the same filtering key in nf_conncount. With the
plain list, we can insert new connections at the tail, so over time
the beginning of list holds long-running connections and those that are
expired, while the newly creates ones are at the end.

Later on we could probably also start to move 'checked ones' to end
of list, so next run has higher chance to find stale entries at the top.

Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Yi-Hung Wei <yihung.wei@gmail.com>